services/utils/auth_utils/miniapp_auth_utils/token_utils.py
from dotenv import load_dotenv
from fastapi import Security, HTTPException
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from pydantic import BaseModel
import jwt
import os

load_dotenv()
JWT_SECRET = os.getenv("JWT_SECRET")
security = HTTPBearer()


class TokenData(BaseModel):
    telegram_id: int

def verify_token(credentials: HTTPAuthorizationCredentials = Security(security)):
    token = credentials.credentials
    print("Token received:", token)
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=["HS256"])
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(401, "Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(401, "Invalid token")