services/utils/auth_utils/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auth_utils"
version = "0.1.0"
description = "Authentication utilities for FastAPI services"
authors = [
    {name = "Avinar"}
]
dependencies = [
    "fastapi>=0.100.0",
    "pydantic>=2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
]